Fix: use fp32 for GGUF gemma — bf16 rejected by GGUF, fp16 by gemma-3
GGUF backend in vLLM forbids bf16; gemma-3 arch forbids fp16 (numerical
instability). Only fp32 satisfies both. Slower + larger KV cache than
bf16, but Q6_K weights stay quantized.

+++ experiments/bloom/bloom_beast_output3.py
@@ -505,9 +505,10 @@ def _vllm_worker_main(req_q, res_q, hf_name: str, gpu_id: int,
             kwargs["tokenizer"] = base
             kwargs["hf_config_path"] = base
             kwargs["quantization"] = "gguf"
-            # GGUF in vLLM only supports fp16/fp32 — auto-pick on Ampere/Hopper
-            # would default to bf16 and crash with a pydantic ValidationError.
-            kwargs["dtype"] = "float16"
+            # vLLM dtype catch-22: GGUF rejects bf16, gemma-3 rejects fp16.
+            # Only fp32 satisfies both. Slower + more KV-cache memory than bf16,
+            # but weights stay Q-quantized so the hit is bearable.
+            kwargs["dtype"] = "float32"
             actual_vocab = LocalModel._inspect_gguf_vocab(gguf_path)
             if actual_vocab is not None:
                 kwargs["hf_overrides"] = {
